requirements.txt
Flask==0.10.1
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.21
configobj==5.0.4
docopt==0.6.1
msgpack-python==0.4.2
path.py==5.1
smartypants==1.8.3
typogrify==2.0.4
watchdog==0.7.1
webassets==0.9
